Add two functions: makeGrid() and changeSize().
makeGrid makes it easier to create a new grid. It is called
once to initialize the grid.

changeSize is to change the number of boxes contained in the grid.
Once changeSize is called, variables needed to create the grid
are re-assigned, all child nodes in the container div are deleted
and a new grid is initialized using makeGrid function.